+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [Unreleased]
 
+### Added
+- Journal entry keyword search from the device browse menu and `/api/journal/search`.
+
 ## [0.1.0] - 2026-05-20
 
 ### Added